% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/readResults.R
\name{read_results}
\alias{read_results}
\alias{readResults}
\title{Read a set of files and create a \code{results_collection}.}
\usage{
read_results(path)

readResults(path)
}
\arguments{
\item{path}{Path to a file or folder with results}
}
\value{
A \code{result_collection} object.
}
\description{
Read a set of files and create a \code{results_collection}.
}
