% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/useAlias.R
\name{use_alias}
\alias{use_alias}
\alias{useAlias}
\title{Function to reexport other functions using camelCase or snake_case.}
\usage{
use_alias(
  original_case = c("camelCase", "snake_case"),
  new_case = c("camelCase", "snake_case")
)

useAlias(
  originalCase = c("camelCase", "snake_case"),
  newCase = c("camelCase", "snake_case")
)
}
\arguments{
\item{originalCase}{Casing of functions to be reexported.}

\item{newCase}{Casing of the new created functions.}
}
\description{
Function to reexport other functions using camelCase or snake_case.
}
